Refactor UE_thread to process symbol by symbol
Samples from radio are read one symbol at a time and they are processed
to produce LLRs for all physical channels. In the last symbol of the
slot, the LLRs are used to decode the physical channel signals.

In the previous commits, OFDM demodulation is performed by each PHY
channel functions. That is now removed because OFDM demod is done
immediately after reading the samples from the radio. For unitary
simulators, a helper function is created that loops over all symbols in
a slot and performs OFDM demodulation.